Fixing compilation

+++ src/data/generators/uniform_generator.cpp
@@ -18,7 +18,7 @@ void UniformGenerator::generate(std::string table_path, std::string query_path){
     // Generate Data
     std::random_device rand_dev;
     std::mt19937 generator(rand_dev());
-    std::uniform_int_distribution<float> distr(0, n_rows);
+    std::uniform_int_distribution<int> distr(0, n_rows);
 
     for(size_t i = 0; i < n_rows; ++i){
         float* row = new float[n_dimensions];
@@ -36,7 +36,7 @@ void UniformGenerator::generate(std::string table_path, std::string query_path){
 
     std::random_device rand_dev_query;
     std::mt19937 generator_query(rand_dev_query());
-    std::uniform_int_distribution<float> distr_query(
+    std::uniform_int_distribution<int> distr_query(
         0, n_rows*(1-per_column_selectivity)
     );
